+++ CHANGELOG.md
@@ -34,3 +34,4 @@
 - [Refator JacksonUtils and JacksonUtilsTest](https://github.com/Tencent/spring-cloud-tencent/pull/365)
 - [refactor:optimize project and code.](https://github.com/Tencent/spring-cloud-tencent/pull/368)
 - [docs: Fix javadoc <br /> error](https://github.com/Tencent/spring-cloud-tencent/pull/371)
+- [UT: add Polaris LoadBalancer unit test](https://github.com/Tencent/spring-cloud-tencent/pull/373)